requirements.txt
pandas
numpy
streamlit
matplotlib
apache-airflow
